+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-## [0.5.1.0b]
+## [0.5.1.0]
 ### Changed
 - data asset folder location in CP4D v4.0.2 (with new git)
 - save Excel/lp/csv files using ibm_watson_studio_lib in CPD v4.0.2 in ScenarioManager